change to double

+++ src/main/java/com/example/springdocker/service/MyMathCalc.java
@@ -12,10 +12,12 @@ public class MyMathCalc {
     public int add(int a, int b){
         return a+b;
     }
+  
     public int multiply(int a, int b){
         return a*b;
     }
-    public float divide(int a, int b){
+  
+    public double divide(int a, int b){
         return a/b;
     }
 }
